refactor(app): move some protocol module inspectors
