+++ app/src/organisms/Devices/ProtocolRun/SetupLiquids/SetupLiquidsMap.tsx
@@ -25,7 +25,7 @@ import { LiquidsLabwareDetailsModal } from './LiquidsLabwareDetailsModal'
 import { getWellFillFromLabwareId } from './utils'
 import { getLabwareRenderInfo } from '../utils/getLabwareRenderInfo'
 import { getStandardDeckViewLayerBlockList } from '/app/local-resources/deck_configuration'
-import { getProtocolModulesInfo } from '../utils/getProtocolModulesInfo'
+import { getProtocolModulesInfo } from '/app/transformations/analysis'
 
 import type {
   CompletedProtocolAnalysis,
