+++ app/src/organisms/Devices/ProtocolRun/utils/__tests__/getLabwareOffsetLocation.test.tsx
@@ -8,7 +8,7 @@ import {
 } from '@opentrons/shared-data'
 import { getLabwareOffsetLocation } from '../getLabwareOffsetLocation'
 import { getLabwareLocation } from '../getLabwareLocation'
-import { getModuleInitialLoadInfo } from '../getModuleInitialLoadInfo'
+import { getModuleInitialLoadInfo } from '/app/transformations/commands'
 import type {
   LoadedLabware,
   LoadedModule,
@@ -17,7 +17,7 @@ import type {
 } from '@opentrons/shared-data'
 
 vi.mock('../getLabwareLocation')
-vi.mock('../getModuleInitialLoadInfo')
+vi.mock('/app/transformations/commands')
 
 const protocolWithTC = (multiple_tipacks_with_tc as unknown) as CompletedProtocolAnalysis
 const mockAdapterDef = opentrons96PcrAdapterV1 as LabwareDefinition2
