+++ app/src/organisms/LabwarePositionCheck/utils/labware.ts
@@ -4,7 +4,7 @@ import {
   getTiprackVolume,
   getLabwareDefURI,
 } from '@opentrons/shared-data'
-import { getModuleInitialLoadInfo } from '../../Devices/ProtocolRun/utils/getModuleInitialLoadInfo'
+import { getModuleInitialLoadInfo } from '/app/transformations/commands'
 import { getLabwareDefinitionsFromCommands } from '/app/molecules/Command/utils/getLabwareDefinitionsFromCommands'
 import type {
   CompletedProtocolAnalysis,
