+++ app/src/organisms/ODD/ProtocolSetup/ProtocolSetupLabware/index.tsx
@@ -38,8 +38,8 @@ import {
 import { FloatingActionButton, SmallButton } from '/app/atoms/buttons'
 import { ODDBackButton } from '/app/molecules/ODDBackButton'
 import { getLabwareSetupItemGroups } from '/app/transformations/commands'
+import { getProtocolModulesInfo } from '/app/transformations/analysis'
 import { useNotifyDeckConfigurationQuery } from '/app/resources/deck_configuration'
-import { getProtocolModulesInfo } from '../../../Devices/ProtocolRun/utils/getProtocolModulesInfo'
 import { getNestedLabwareInfo } from '../../../Devices/ProtocolRun/SetupLabware/getNestedLabwareInfo'
 import { LabwareStackModal } from '../../../Devices/ProtocolRun/SetupLabware/LabwareStackModal'
 import { useMostRecentCompletedAnalysis } from '../../../LabwarePositionCheck/useMostRecentCompletedAnalysis'
