+++ app/src/transformations/analysis/__tests__/getProtocolModulesInfo.test.ts
