+++ app/src/transformations/analysis/index.ts
@@ -1,1 +1,2 @@
 export * from './parseProtocolAnalysisOutput'
+export * from './getProtocolModulesInfo'
