+++ app/src/transformations/commands/transformations/getModuleInitialLoadInfo.ts
