+++ app/src/transformations/commands/transformations/index.ts
@@ -2,3 +2,4 @@ export * from './getLabwareSetupItemGroups'
 export * from './getProtocolUsesGripper'
 export * from './formatDuration'
 export * from './formatInterval'
+export * from './getModuleInitialLoadInfo'
